+++ Source/SpaceHorror/SpaceHorrorCharacter.h
@@ -45,13 +45,9 @@ class ASpaceHorrorCharacter : public ACharacter
 	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = "true"))
 	class UMotionControllerComponent* L_MotionController;
 	
-	/** CT */
-	UPROPERTY(VisibleDefaultsOnly, Category = Mesh)
-	class USkeletalMeshComponent* DC_Gun;
+	UPROPERTY(VisibleDefaultsOnly, Category = Weapon)
+	class UChildActorComponent* DC_Gun = nullptr;
 
-	/** CT */
-	UPROPERTY(VisibleDefaultsOnly, Category = Mesh)
-	class USceneComponent* DC_MuzzleLocation;
 
 public:
 	ASpaceHorrorCharacter();
@@ -181,8 +177,8 @@ private:
 	void Weapon4();
 	void Weapon5();
 	void Weapon6();
-
-	void CreateCustomGun(FName gunname, USkeletalMeshComponent* gunmesh, USceneComponent* muzzlelocation);
 	
+
+
 };